refactor(core): move NetworkMonitor and update subscription business
- Moved `NetworkMonitor` interface to the `core/domain` module.
- Updated `AppModule` to use the new `NetworkMonitor` location.
- Updated `SubscriptionsBusinessRepositoryImpl` to use `Subscription` model and its extension function to create `SubscriptionRequest`
- Updated `Invoice` model to use the new extension function for date formatting.
- Removed the old `NetworkMonitor` interface from the `core/data` module.
