+++ core/data/src/main/java/com/casecode/pos/core/data/utils/ConnectivityManagerNetworkMonitor.kt
@@ -23,6 +23,7 @@ import android.net.NetworkCapabilities
 import android.net.NetworkRequest
 import android.net.NetworkRequest.Builder
 import androidx.core.content.getSystemService
+import com.casecode.pos.core.domain.utils.NetworkMonitor
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.channels.awaitClose
