+++ core/data/src/main/java/com/casecode/pos/core/data/utils/NetworkMonitor.kt
@@ -15,8 +15,3 @@
  */
 package com.casecode.pos.core.data.utils
 
-import kotlinx.coroutines.flow.Flow
-
-interface NetworkMonitor {
-    val isOnline: Flow<Boolean>
-}
